fix: move typings into src and rename *.ts -> *.d.ts Typings weren't included in npm package causing `Cannot find module '../../typings/utils'.`
