+++ lib/src/DatePicker/DatePicker.d.ts
@@ -1,8 +1,8 @@
 import { ComponentClass, ReactNode } from 'react';
 import { DateType } from '../constants/prop-types';
-import { Utils } from '../../typings/utils';
+import { Utils } from '../typings/utils';
 import { RenderDay } from './Calendar';
-import { MaterialUiPickersDate } from '../../typings/date'
+import { MaterialUiPickersDate } from '../typings/date'
 
 export interface DatePickerProps {
     date: MaterialUiPickersDate;
