+++ lib/src/DatePicker/YearSelection.d.ts
@@ -1,7 +1,7 @@
 import { ComponentClass, ReactNode } from 'react';
 import { DateType } from '../constants/prop-types';
-import { Utils } from '../../typings/utils';
-import { MaterialUiPickersDate } from '../../typings/date'
+import { Utils } from '../typings/utils';
+import { MaterialUiPickersDate } from '../typings/date'
 
 export interface YearSelectionProps {
     date: MaterialUiPickersDate;
