+++ lib/src/DatePicker/index.d.ts
@@ -1,9 +1,9 @@
 import { ComponentClass, ReactNode } from 'react';
 import { DateType } from '../constants/prop-types';
-import { Utils } from '../../typings/utils';
+import { Utils } from '../typings/utils';
 import { RenderDay } from './Calendar';
 import { ModalWrapperProps } from '../wrappers/ModalWrapper';
-import { MaterialUiPickersDate } from '../../typings/date'
+import { MaterialUiPickersDate } from '../typings/date'
 import { PickerBaseProps } from '../_shared/PickerBase'
 import { Omit } from 'material-ui'
