+++ lib/src/DateTimePicker/DateTimePickerHeader.d.ts
@@ -1,7 +1,7 @@
 import { ComponentClass } from 'react';
 import { DateTimePickerView } from '../constants/date-picker-view';
-import { Utils } from '../../typings/utils';
-import { MaterialUiPickersDate } from '../../typings/date'
+import { Utils } from '../typings/utils';
+import { MaterialUiPickersDate } from '../typings/date'
 
 type MeridiemMode = 'am' | 'pm';
