+++ lib/src/DateTimePicker/index.d.ts
@@ -1,10 +1,10 @@
 import { ComponentClass, ReactNode } from 'react';
 import { DateTimePickerView } from '../constants/date-picker-view';
 import { DateType } from '../constants/prop-types';
-import { Utils } from '../../typings/utils';
+import { Utils } from '../typings/utils';
 import { RenderDay } from '../DatePicker/Calendar';
 import { ModalWrapperProps } from '../wrappers/ModalWrapper';
-import { MaterialUiPickersDate } from '../../typings/date'
+import { MaterialUiPickersDate } from '../typings/date'
 import { PickerBaseProps } from '../_shared/PickerBase'
 import { Omit } from 'material-ui'
