+++ lib/src/utils/moment-utils.d.ts
@@ -1,4 +1,4 @@
-import { Utils } from '../../typings/utils'
+import { Utils } from '../typings/utils'
 
 declare const MomentUtils: Utils;
 export default MomentUtils
